+++ R/frame-data.R
@@ -58,17 +58,15 @@ frame_data <- function(...) {
     ))
   }
 
-  frame_nrow <- n_elements / frame_ncol
-
-  # Extract the columns from 'frame_rest'
-  frame_columns <- lapply(seq_len(frame_ncol), function(i) {
-    indices <- seq.default(from = i, to = length(frame_rest), by = frame_ncol)
-    unlist(frame_rest[indices])
+  frame_mat <- matrix(frame_rest, ncol = frame_ncol, byrow = TRUE)
+  frame_col <- lapply(seq_len(ncol(frame_mat)), function(i) {
+    col <- frame_mat[, i]
+    if (is.list(col[[1]])) col else unlist(col)
   })
 
   # Create a tbl_df and return it
-  names(frame_columns) <- frame_names
-  as_data_frame(frame_columns)
+  names(frame_col) <- frame_names
+  as_data_frame(frame_col)
 }
 
 #' @rdname frame_data