+++ inst/include/dplyr/Gatherer.h
@@ -18,15 +18,18 @@ namespace dplyr {
         GathererImpl( RObject& first, SlicingIndex& indices, Proxy& proxy_, const Data& gdf_, int first_non_na_ ) :
             gdf(gdf_), proxy(proxy_), data(gdf.nrows(), Vector<RTYPE>::get_na() ), first_non_na(first_non_na_)
         {
-            grab( first, indices ) ;
+            if( first_non_na < gdf.ngroups() )
+              grab( first, indices ) ;
             copy_most_attributes( data, first ) ;
         }
 
         SEXP collect(){
             int ngroups = gdf.ngroups() ;
+            if( first_non_na == ngroups ) return data ;
             typename Data::group_iterator git = gdf.group_begin() ;
             int i = 0 ;
             for(; i<first_non_na; i++) ++git ;
+            ++git; i++ ;
             for(; i<ngroups; i++, ++git){
                 SlicingIndex indices = *git ;
                 Shield<SEXP> subset( proxy.get( indices ) ) ;
@@ -37,16 +40,16 @@ namespace dplyr {
 
     private:
 
-        inline void grab(SEXP data, const SlicingIndex& indices){
-            int n = Rf_length(data) ;
-            if( is<LogicalVector>(data) && all(is_na(LogicalVector(data))).is_true() ){
+        inline void grab(SEXP subset, const SlicingIndex& indices){
+            int n = Rf_length(subset) ;
+            if( is<LogicalVector>(subset) && all(is_na(LogicalVector(subset))).is_true() ){
               grab_rep( Vector<RTYPE>::get_na(), indices ) ;
             } else {
-              check_type(data) ;
+              check_type(subset) ;
               if(n == indices.size() ){
-                  grab_along( data, indices ) ;
+                  grab_along( subset, indices ) ;
               } else if( n == 1) {
-                  grab_rep( Rcpp::internal::r_vector_start<RTYPE>(data)[0], indices ) ;
+                  grab_rep( Rcpp::internal::r_vector_start<RTYPE>(subset)[0], indices ) ;
               } else {
                   stop ( "incompatible size (%d), expecting %d (the group size) or 1",
                           n, indices.size()) ;
@@ -83,15 +86,97 @@ namespace dplyr {
     } ;
 
     template <typename Data, typename Subsets>
+    class ListGatherer : public Gatherer {
+    public:
+        typedef GroupedCallProxy<Data,Subsets> Proxy ;
+
+        ListGatherer( List first, SlicingIndex& indices, Proxy& proxy_, const Data& gdf_, int first_non_na_ ) :
+            gdf(gdf_), proxy(proxy_), data(gdf.nrows()), first_non_na(first_non_na_)
+        {
+            if( first_non_na < gdf.ngroups() ){
+                perhaps_duplicate(first) ;
+                grab( first, indices ) ;
+            }
+
+            copy_most_attributes( data, first ) ;
+        }
+
+        SEXP collect(){
+            int ngroups = gdf.ngroups() ;
+            if( first_non_na == ngroups ) return data ;
+            typename Data::group_iterator git = gdf.group_begin() ;
+            int i = 0 ;
+            for(; i<first_non_na; i++) ++git ;
+            ++git; i++ ;
+            for(; i<ngroups; i++, ++git){
+                SlicingIndex indices = *git ;
+                List subset( proxy.get(indices) ) ;
+                perhaps_duplicate(subset) ;
+                grab(subset, indices);
+            }
+            return data ;
+        }
+
+    private:
+
+        inline void perhaps_duplicate( List& x ){
+          int n = x.size() ;
+          for( int i=0; i<n; i++){
+              SEXP xi = x[i] ;
+              if( IS_DPLYR_SHRINKABLE_VECTOR(xi) ) {
+                x[i] = Rf_duplicate(xi) ;
+              } else if( TYPEOF(xi) == VECSXP ){
+                List lxi(xi) ;
+                perhaps_duplicate( lxi ) ;
+              }
+          }
+        }
+
+        inline void grab(const List& subset, const SlicingIndex& indices){
+            int n = subset.size() ;
+
+            if(n == indices.size() ){
+                grab_along( subset, indices ) ;
+            } else if( n == 1) {
+                grab_rep( subset[0], indices ) ;
+            } else {
+                stop ( "incompatible size (%d), expecting %d (the group size) or 1",
+                        n, indices.size()) ;
+            }
+        }
+
+        void grab_along( const List& subset, const SlicingIndex& indices ){
+            int n = indices.size();
+            for( int j=0; j<n; j++){
+                data[ indices[j] ] = subset[j] ;
+            }
+        }
+
+        void grab_rep( SEXP value, const SlicingIndex& indices ){
+            int n = indices.size();
+            for( int j=0; j<n; j++){
+                data[ indices[j] ] = value ;
+            }
+        }
+
+        const Data& gdf ;
+        Proxy& proxy ;
+        List data ;
+        int first_non_na ;
+
+    } ;
+
+    template <typename Data, typename Subsets>
     class FactorGatherer : public Gatherer {
     public:
         typedef GroupedCallProxy<Data,Subsets> Proxy ;
         typedef IntegerVector Factor;
 
         FactorGatherer( RObject& first, SlicingIndex& indices, Proxy& proxy_, const Data& gdf_, int first_non_na_ ) :
-          levels(), data(gdf_.nrows()), first_non_na(first_non_na_), proxy(proxy_), gdf(gdf_)
+          levels(), data(gdf_.nrows(), NA_INTEGER), first_non_na(first_non_na_), proxy(proxy_), gdf(gdf_)
         {
-          grab( (SEXP)first, indices ) ;
+          if( first_non_na <  gdf.ngroups() )
+            grab( (SEXP)first, indices ) ;
           copy_most_attributes( data, first ) ;
         }
 
@@ -137,13 +222,24 @@ namespace dplyr {
 
             // grab data
             int n = indices.size() ;
-            for( int i=0; i<n; i++){
-                if( f[i] == NA_INTEGER ){
-                  data[ indices[i] ] = NA_INTEGER ;
-                } else {
-                  data[ indices[i] ] = matches[ f[i] - 1 ] ;
-                }
 
+            int nf = f.size() ;
+            if( n == nf ){
+              for( int i=0; i<n; i++){
+                  if( f[i] != NA_INTEGER ){
+                    data[ indices[i] ] = matches[ f[i] - 1 ] ;
+                  }
+              }
+            } else if( nf == 1){
+              int value = NA_INTEGER  ;
+              if( f[0] != NA_INTEGER ){
+                value = matches[ f[0] - 1] ;
+                for( int i=0; i<n; i++){
+                  data[ indices[i] ] = value ;
+                }
+              }
+            } else {
+              stop( "incompatible size" ) ;
             }
         }
 
@@ -188,12 +284,15 @@ namespace dplyr {
         typename Data::group_iterator git = gdf.group_begin() ;
         SlicingIndex indices = *git ;
         RObject first( proxy.get(indices) ) ;
+
         if( Rf_inherits(first, "POSIXlt" ) ){
             stop("`mutate` does not support `POSIXlt` results");
         }
         int ng = gdf.ngroups() ;
-        int i = 1 ; ++git ;
-        for( ; all_na(first) && i<ng; i++, ++git){
+        int i = 0 ;
+        while( all_na(first) ){
+          i++ ; if( i == ng ) break ;
+          ++git ;
           indices = *git ;
           first = proxy.get(indices) ;
         }
@@ -208,7 +307,7 @@ namespace dplyr {
             case REALSXP: return new GathererImpl<REALSXP,Data,Subsets> ( first, indices, proxy, gdf, i ) ;
             case LGLSXP:  return new GathererImpl<LGLSXP,Data,Subsets>  ( first, indices, proxy, gdf, i ) ;
             case STRSXP:  return new GathererImpl<STRSXP,Data,Subsets>  ( first, indices, proxy, gdf, i ) ;
-            case VECSXP:  return new GathererImpl<VECSXP,Data,Subsets>  ( first, indices, proxy, gdf, i ) ;
+            case VECSXP:  return new ListGatherer<Data,Subsets>  ( List(first), indices, proxy, gdf, i ) ;
             case CPLXSXP: return new GathererImpl<CPLXSXP,Data,Subsets> ( first, indices, proxy, gdf, i ) ;
             default: break ;
         }